requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
Django==3.1.4
django-webpack-loader==0.7.0
djangorestframework==3.12.2
flake8==3.8.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
